# Improves performance of json responses